C#: Add buffer pool

+++ csharp/TinyNF.Environment/IEnvironment.cs
@@ -5,8 +5,9 @@ namespace TinyNF.Environment
     public interface IEnvironment
     {
         // Memory
-        Memory<T> Allocate<T>(nuint count) where T : struct; // The semantics of this are rather silly: the resulting Memory<T> has an infinite lifetime!
-        Memory<T> MapPhysicalMemory<T>(ulong addr, uint count) where T : unmanaged; // addr is ulong, not nuint, because PCI BARs are 64-bit
+        // TODO this is dangerous, we need an initializer parameter
+        Memory<T> Allocate<T>(int count) where T : struct; // The semantics of this are rather silly: the resulting Memory<T> has an infinite lifetime!
+        Memory<T> MapPhysicalMemory<T>(ulong addr, int count) where T : unmanaged; // addr is ulong, not nuint, because PCI BARs are 64-bit
         nuint GetPhysicalAddress<T>(ref T value);
 
         // PCI
